Created the README.md file
I Created the README.md file, because GitHub suggested that I do so. I was gonna do it later and actually add some more information. but I decided to go ahead and just create the file for now. and add more information to it later. though I did add all the legalalitys that are needed for this repository